requirements.txt
Pillow
tk
customtkinter
diffusers
transformers
accelerate